+++ constants.js
@@ -1,0 +1,23 @@
+export const loanProfileTypesMap = {
+  FIXED: '1',
+  ROLLING: '2',
+  INDEFINITE: '3',
+};
+
+export const intervalPeriodsMap = {
+  MINUTE: 'minutes',
+  HOUR: 'hours',
+  DAY: 'days',
+  WEEK: 'weeks',
+  MONTH: 'months',
+};
+
+export const intervalIdsMap = {
+  1: 'minutes',
+  2: 'hours',
+  3: 'days',
+  4: 'weeks',
+  5: 'months',
+};
+
+export default {};