+++ CHANGELOG.md
@@ -36,6 +36,12 @@
 * Additional "No proxies/No sponsors" formatting. Refs UX-115.
 * Rewire loan links from items to inventory. Fixes UIU-368.
 * Relabel elements in user details Proxy section. Fixes UIU-370.
+* Get fixed renewal period from loan policy. Fixes UIU-405.
+* Get rolling renewal period from loan policy. Fixes UIU-415.
+* Add renew button to loan details. Fixes UIU-395.
+* Make perm set title required. Fixes UIU-412.
+* Fix address validation. Fixes UIU-414.
+* Add placeholder text to explain what's wrong when there is no "find-user" plugin. Fixes UIU-421.
 
 ## [2.12.0](https://github.com/folio-org/ui-users/tree/v2.12.0) (2017-11-28)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.11.0...v2.12.0)